+++ purchase/src/main/java/com/example/purchase/domain/PurchaseProduct.kt
@@ -2,6 +2,7 @@ package com.example.purchase.domain
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
-data class PurchaseProduct(val name: String, val price: Int): Parcelable
+data class PurchaseProduct(val name: String, val price: BigDecimal): Parcelable
